fixes after refactor
