+++ yarn-project/acir-simulator/src/avm/index.test.ts
@@ -40,7 +40,6 @@ describe('avm', () => {
 
     // Initialize AVM context
     const context = new AvmContext(journal, initExecutionEnvironment({ calldata }));
-    await context.fetchAndDecodeBytecode();
 
     // Execute AVM
     const results = await context.execute();
@@ -67,7 +66,6 @@ describe('avm', () => {
 
       // Initialize AVM context
       const context = new AvmContext(journal, initExecutionEnvironment({ calldata }));
-      await context.fetchAndDecodeBytecode();
 
       // Execute AVM
       const results = await context.execute();
